fix: fix packet

+++ src/mc/codebuilder/Category.h
@@ -4,5 +4,9 @@
 
 namespace CodeBuilderStorageQueryOptions {
 
-enum class Category : int { None = 0x0, CodeStatus = 0x1, Instantiation = 0x2 };
+enum class Category : uchar {
+    None          = 0x0,
+    CodeStatus    = 0x1,
+    Instantiation = 0x2,
+};
 };
