+++ src/mc/codebuilder/Operation.h
@@ -4,5 +4,10 @@
 
 namespace CodeBuilderStorageQueryOptions {
 
-enum class Operation : int { None = 0x0, Get = 0x1, Set = 0x2, Reset = 0x3 };
+enum class Operation : uchar {
+    None  = 0x0,
+    Get   = 0x1,
+    Set   = 0x2,
+    Reset = 0x3,
+};
 };
